Trocado int para float e adicionado .2f para duas casas decimais

+++ media_nota.c
@@ -2,23 +2,23 @@
 
 int main()
 {
-    int nota1, nota2, nota3;
-    int media;
+    float nota1, nota2, nota3;
+    float media;
 
     printf("***Programa de cálculo de média***\n");
 
     printf("Digite sua primeira nota:\n");
-    scanf("%d", &nota1);
+    scanf("%f", &nota1);
 
     printf("Digite sua segunda nota:\n");
-    scanf("%d", &nota2);
+    scanf("%f", &nota2);
 
     printf("Digite sua  terceira nota:\n");
-    scanf("%d", &nota3);
+    scanf("%f", &nota3);
 
     media = (nota1+nota2+nota3)/3;
 
-    printf("A média é %d\n", media);
+    printf("A média é %.2f\n", media);
 
     return 0;
 }
